Move Polaris components into their own files
This makes importing those files cleaner.
